* modules/gui/wxwindows/*: added skeleton for popup contextual menu. * modules/access/dvdplay/access.c, modules/audio_filter/resampler/linear.c: a couple of warning fixes.

+++ modules/access/dvdplay/access.c
@@ -2,7 +2,7 @@
  * access.c: access capabilities for dvdplay plugin.
  *****************************************************************************
  * Copyright (C) 2001 VideoLAN
- * $Id: access.c,v 1.7 2002/12/06 16:34:04 sam Exp $
+ * $Id: access.c,v 1.8 2002/12/13 01:50:32 gbazin Exp $
  *
  * Author: Stéphane Borel <stef@via.ecp.fr>
  *
@@ -301,7 +301,7 @@ static int dvdplay_SetArea( input_thread_t * p_input, input_area_t * p_area )
      * Chapter selection
      */
 
-    if( p_area->i_part != dvdplay_chapter_cur( p_dvd->vmg ) )
+    if( (int)p_area->i_part != dvdplay_chapter_cur( p_dvd->vmg ) )
     {
         if( ( p_area->i_part > 0 ) &&
             ( p_area->i_part <= p_area->i_part_nb ))
